Added `photodiodeTest.m`
An oscilloscope and photodiode/transistor are needed for this code to
be useful.

This code checks refresh rate by flashing a white square in the
top-left corner for one frame out of every `testFrames` (the refresh
rate you want to test).

If the oscilloscope reports a single peak in a second, your monitor is
running at the refresh rate you think it is. If not, your refresh rate
is not set correctly.